Styling Styling

+++ src/IndividualArticleServlet.java
@@ -32,13 +32,6 @@ public class IndividualArticleServlet extends HttpServlet {
         String csrfSessionToken = (String)session.getAttribute("csrfSessionToken");
         String csrfToken = req.getParameter("csrfToken");
 
-        if (!csrfSessionToken.equals(csrfToken)){
-            System.out.println("csrfTokens not verified");
-            resp.sendError(666);
-            return;
-        }
-
-        System.out.println("csrfTokens verified");
         ServletContext servletContext = getServletContext();
         String sessionFilePath = servletContext.getRealPath("WEB-INF/Sessions");
         String sessionID = session.getId();
@@ -77,6 +70,11 @@ public class IndividualArticleServlet extends HttpServlet {
 
             //add articles
             if ("add".equals(op)) {
+                if(!passTokenCheck(csrfToken, csrfSessionToken)){
+                    resp.sendError(666);
+                    return;
+                }
+
                 String title = req.getParameter("title");
                 String content = req.getParameter("content");
                 user = String.valueOf( session.getAttribute("personLoggedIn"));
@@ -98,6 +96,10 @@ public class IndividualArticleServlet extends HttpServlet {
 //                adding articles goes to my articles
                 req.getRequestDispatcher("myArticles").forward(req, resp);
             } else if ("delete".equals(op)) {
+                if(!passTokenCheck(csrfToken, csrfSessionToken)){
+                    resp.sendError(666);
+                    return;
+                }
                 System.out.println("IndividualArticleServlet: Delete option");
                 System.out.println(req.getParameter("articleID"));
                 int id = Integer.parseInt(req.getParameter("articleID"));
@@ -108,24 +110,40 @@ public class IndividualArticleServlet extends HttpServlet {
                 System.out.println("IndividualArticleServlet: Else Statement");
                 //deals with article manipulation
                 if ("commentOnArticle".equals(op)) {
+                    if(!passTokenCheck(csrfToken, csrfSessionToken)){
+                        resp.sendError(666);
+                        return;
+                    }
                     String userWhoCommented = req.getParameter("userWhoCommented");
                     String comment = req.getParameter("newComment");
                 int articleID = Integer.parseInt(req.getParameter("articleID"));
                     java.sql.Date sqlDate = java.sql.Date.valueOf(LocalDate.now());
                     dao.addCommentToArticle(articleID, userWhoCommented, sqlDate, comment);
                 } else if ("editarticle".equals(op)) {
+                    if(!passTokenCheck(csrfToken, csrfSessionToken)){
+                        resp.sendError(666);
+                        return;
+                    }
                     String title = req.getParameter("title");
                     String content = req.getParameter("content");
                     java.sql.Date sqlDate = java.sql.Date.valueOf(LocalDate.now());
                     dao.addArticle(title, content, user, sqlDate);
                     System.out.println("IndividualArticleServlet: new article made");
                 } else if ("deleteCommentOnArticle".equals(op)) {
+                    if(!passTokenCheck(csrfToken, csrfSessionToken)){
+                        resp.sendError(666);
+                        return;
+                    }
                     System.out.println("IndividualArticleServlet: delete button pressed");
                     String commentIDToBeDeleted = req.getParameter("commentID");
                     System.out.println();
                     System.out.println("IAS: id - " + commentIDToBeDeleted);
                     dao.deleteCommentOnArticle(Integer.parseInt(commentIDToBeDeleted));
                 } else if ("replyToAComment".equals(op)) {
+                    if(!passTokenCheck(csrfToken, csrfSessionToken)){
+                        resp.sendError(666);
+                        return;
+                    }
                     System.out.println("IndividualArticleServlet replying to comment button pressed");
                     java.sql.Date sqlDate = java.sql.Date.valueOf(LocalDate.now());
                     String userWhoCommented = req.getParameter("userWhoCommented");
@@ -154,5 +172,15 @@ public class IndividualArticleServlet extends HttpServlet {
         }
 
         }
+    }
+
+    private boolean passTokenCheck(String token, String sessionToken){
+        if (!sessionToken.equals(token)) {
+            System.out.println("csrfTokens not verified");
+            return false;
+        }
+
+        System.out.println("token verification passed!");
+        return true;
     }
 }
